requirements.txt
Flask==3.1.0
icalendar==6.1.0
python-dateutil==2.9.0
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
requests==2.32.3
pytz==2024.2
